+++ src/app.js
@@ -1,0 +1,13 @@
+import login        from './views/pages/login.js'
+import register     from './views/pages/register.js'
+
+
+const routes = {
+'/login'      : login
+'/register'   : register
+
+}; 
+
+
+
+window.addEventListener('hashchange', router);
