+++ src/index.html
@@ -21,12 +21,13 @@
 
 
     <div class="down">
-    <button id="btn1">REGISTRARSE</button>
-    <button id="btn2">INICIAR SESIÓN</button>
+    <a href = "#btn1">REGISTRARSE</a>
+   <a href = "#btn2">INICIAR SESIÓN</a>
     </div>
     
-
     
+
+      
 
 
   <script src="data.js"></script>
