+++ src/main.js
@@ -8,48 +8,31 @@ const buttonAccess = document.getElementById('access');
 
 
 //FUNCION REGISTRAR USUARIO
-buttonRegister.addEventListener('click', () => {
 
-let email = document.getElementById('email').value;
-let pass = document.getElementById('pass').value;
 
-firebase.auth().createUserWithEmailAndPassword(email, pass).then((res)=>{
-  console.log(res)
-})
-.catch(function(error) {
-    // Handle Errors here.
-    var errorCode = error.code;
-    var errorMessage = error.message;
-    // ...
 
-  console.log(errorCode);
-  console.log(errorMessage);
-
-  });
-
-});
 
 //FUNCION INGRESA USUARIO
-buttonAccess.addEventListener('click', () => {
+//buttonAccess.addEventListener('click', () => {
 
-    let emailAc = document.getElementById('email-ac').value;
-    let passAc = document.getElementById('pass-ac').value;
+    //let emailAc = document.getElementById('email-ac').value;
+    //let passAc = document.getElementById('pass-ac').value;
     
-    firebase.auth().signInWithEmailAndPassword(emailAc, passAc).then((res)=>{
-      console.log(res);
-    })
-    .catch(function(error) {
+    //firebase.auth().signInWithEmailAndPassword(emailAc, passAc).then((res)=>{
+      //console.log(res);
+    //})
+    //.catch(function(error) {
         // Handle Errors here.
-        var errorCode = error.code;
-        var errorMessage = error.message;
+        //var errorCode = error.code;
+        //var errorMessage = error.message;
         // ...
     
-    console.log(errorCode);
-    console.log(errorMessage);
+    //console.log(errorCode);
+    //console.log(errorMessage);
     
-      });
+      //});
     
-    });
+    //});
 
 //VERIFICAR USUARIO
 const verify = () => {
