--- src/registro.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel = "stylesheet" href="style.css"/>
-    <title>Document</title>
-    <script src="https://www.gstatic.com/firebasejs/5.8.5/firebase.js"></script>
-    <script>
-      // Initialize Firebase
-      var config = {
-        apiKey: "AIzaSyDOvN36a9Ec3svVs5XpAwUF0gd9wBaVy7c",
-        authDomain: "social-network-f8e7a.firebaseapp.com",
-        databaseURL: "https://social-network-f8e7a.firebaseio.com",
-        projectId: "social-network-f8e7a",
-        storageBucket: "social-network-f8e7a.appspot.com",
-        messagingSenderId: "454382512914"
-      };
-      firebase.initializeApp(config);
-    </script>
-</head>
-<body class="registration">
-    <!--registrarse-->
-<h4>Registro de Usuarios</h4>
-<input id="email" type ="email" placeholder="Ingresa email">
-<input id="pass" type ="password" placeholder="Ingresa password">
-<!-- <button id="register" onclick="register()">Enviar</button> -->
-
-
-
-<section id="user-data">
-        
-        </section>
-        <script src="data.js"></script>
-        <script src="main.js"></script> 
-</body>
-</html>
